requirements.txt
google-ads
pandas
openpyxl
python-dotenv
protobuf~=4.25.8